+++ src/main/java/com/example/pro/mapper/EmpMapper.java
@@ -1,0 +1,15 @@
+package com.example.pro.mapper;
+
+import org.apache.ibatis.annotations.Delete;
+
+public interface EmpMapper {
+
+    /**
+     * 删除
+     * @param empno
+     * @return
+     */
+    @Delete("delete from emp where empno=#{empno}")
+    int del(Integer empno);
+
+}
