--- src/main/java/controller/EmpController.java
@@ -1,30 +1,0 @@
-package controller;
-
-import org.springframework.web.bind.annotation.RequestMapping;
-import service.EmpService;
-
-@RequestMapping
-public class EmpController {
-
-    private EmpService empService;
-
-
-    //删除
-
-
-
-
-    //修改
-
-
-
-
-    //查询
-
-
-
-
-    //新增
-
-
-}
