--- src/main/java/mapper/EmpMapper.java
@@ -1,24 +1,0 @@
-package mapper;
-
-
-public interface EmpMapper {
-
-
-    //删除
-
-
-
-
-    //修改
-
-
-
-
-    //查询
-
-
-
-    //新增
-
-
-}
