--- src/main/java/service/EmpService.java
@@ -1,31 +1,0 @@
-package service;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-@Service
-public class EmpService {
-
-    @Autowired
-    private EmpMapper empMapper;
-
-
-    //删除
-
-
-
-
-    //修改
-
-
-
-
-    //查询
-
-
-
-
-    //新增
-
-
-}
