+++ src/main/resources/application.yml
@@ -1,10 +1,10 @@
 server:
-port: 8080
-spring:
-datasource:
-password: root
-username: root
+  port: 8080
+  spring:
+  datasource:
+    password: root
+    username: root
 url: jdbc:mysql://localhost/sys?useLL=false&serverTimezone=Asia/Shanghai&useUnicode=true&characterEncoding=utf8&useSSL=true
-mybatis: #\u914D\u7F6E\u522B\u540D
-type-aliases-package: com.example.pro.domain
-mapper-locations: classpath:com/example/pro/mapper/*.xml
+  mybatis: #\u914D\u7F6E\u522B\u540D
+  type-aliases-package: com.example.pro.domain
+  mapper-locations: classpath:com/example/pro/mapper/*.xml
